# packages/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Jobseek dashboard API
  version: 0.1.0
  description: Versioned private dashboard API. All data routes require authentication.
servers:
  - url: /api/v1
paths:
  /interviews/prepare:
    post:
      operationId: prepareInterview
      summary: Commission one interview brief from complete owner supplied context
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          { application/json: { schema: { $ref: '#/components/schemas/PrepareInterviewRequest' } } }
      responses:
        '200':
          {
            description: Existing commission,
            content:
              {
                application/json:
                  { schema: { $ref: '#/components/schemas/InterviewCommissionResponse' } },
              },
          }
        '201':
          {
            description: Commission created,
            content:
              {
                application/json:
                  { schema: { $ref: '#/components/schemas/InterviewCommissionResponse' } },
              },
          }
        '409': { $ref: '#/components/responses/Conflict' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /interviews:
    get:
      operationId: listInterviews
      summary: List private owner interview preparations
      security: [{ cookieAuth: [] }]
      responses:
        '200':
          {
            description: Interview list,
            content:
              { application/json: { schema: { $ref: '#/components/schemas/InterviewList' } } },
          }
  /interviews/{id}:
    get:
      operationId: getInterview
      summary: Read one sourced interview brief and owner debriefs
      security: [{ cookieAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          {
            description: Interview and debriefs,
            content:
              { application/json: { schema: { $ref: '#/components/schemas/InterviewDetail' } } },
          }
        '404': { $ref: '#/components/responses/NotFound' }
  /interviews/{id}/debrief:
    post:
      operationId: debriefInterview
      summary: Commission a cited owner reported debrief from complete notes
      security: [{ cookieAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          { application/json: { schema: { $ref: '#/components/schemas/DebriefInterviewRequest' } } }
      responses:
        '200':
          {
            description: Existing debrief commission,
            content:
              {
                application/json:
                  { schema: { $ref: '#/components/schemas/DebriefInterviewResponse' } },
              },
          }
        '201':
          {
            description: Debrief commission created,
            content:
              {
                application/json:
                  { schema: { $ref: '#/components/schemas/DebriefInterviewResponse' } },
              },
          }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /rounds/active:
    get:
      operationId: getActiveRound
      summary: Read the active commissioned round
      security: [{ cookieAuth: [] }]
      responses:
        '200':
          {
            description: Active round,
            content: { application/json: { schema: { $ref: '#/components/schemas/Round' } } },
          }
        '404': { $ref: '#/components/responses/NotFound' }
  /correspondence/threads:
    get:
      operationId: listCorrespondenceThreads
      summary: List private mirrored correspondence threads
      security: [{ cookieAuth: [] }]
      responses:
        '200':
          {
            description: Thread list,
            content:
              {
                application/json:
                  { schema: { $ref: '#/components/schemas/CorrespondenceThreadList' } },
              },
          }
  /correspondence/threads/{id}:
    get:
      operationId: getCorrespondenceThread
      summary: Read one mirrored thread with its complete saved messages
      security: [{ cookieAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          {
            description: Thread and messages,
            content:
              {
                application/json:
                  { schema: { $ref: '#/components/schemas/CorrespondenceThreadDetail' } },
              },
          }
        '404': { $ref: '#/components/responses/NotFound' }
  /correspondence/import:
    post:
      operationId: importCorrespondenceThreads
      summary: Import owner-supplied pasted threads into the read-only mirror
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          {
            application/json:
              { schema: { $ref: '#/components/schemas/ImportCorrespondenceRequest' } },
          }
      responses:
        '200':
          {
            description: Import result,
            content:
              {
                application/json:
                  { schema: { $ref: '#/components/schemas/ImportCorrespondenceResponse' } },
              },
          }
  /correspondence/threads/{id}/process:
    post:
      operationId: processCorrespondenceThread
      summary: Commission reply interpretation and a cited follow-up draft for one thread
      security: [{ cookieAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          { application/json: { schema: { $ref: '#/components/schemas/ProcessRepliesRequest' } } }
      responses:
        '200':
          {
            description: Existing commission,
            content:
              {
                application/json:
                  { schema: { $ref: '#/components/schemas/ReplyCommissionResponse' } },
              },
          }
        '201':
          {
            description: Commission created,
            content:
              {
                application/json:
                  { schema: { $ref: '#/components/schemas/ReplyCommissionResponse' } },
              },
          }
        '409': { $ref: '#/components/responses/Conflict' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /replies/{id}:
    get:
      operationId: getReplyProcessing
      summary: Read one reply processing with its saved intent and draft
      security: [{ cookieAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          {
            description: Processing and draft,
            content:
              {
                application/json:
                  { schema: { $ref: '#/components/schemas/ReplyProcessingDetail' } },
              },
          }
        '404': { $ref: '#/components/responses/NotFound' }
  /rounds/latest-completed:
    get:
      operationId: getLatestCompletedRound
      summary: Read the owner's latest saved round result, including recorded partial delivery
      security: [{ cookieAuth: [] }]
      parameters:
        - in: query
          name: outcome
          required: true
          schema:
            {
              type: string,
              enum:
                [
                  all,
                  process_input,
                  prepare,
                  compare_offers,
                  deliver,
                  interview_prepare,
                  interview_debrief,
                  process_replies,
                ],
            }
      responses:
        '200':
          description: Latest matching saved result with transient next-action currentness, or null when none exists
          content:
            application/json:
              schema:
                allOf: [{ $ref: '#/components/schemas/Round' }]
                nullable: true
        '400': { $ref: '#/components/responses/ValidationError' }
  /rounds/prepare:
    post:
      operationId: prepareRound
      summary: Owner commissions a private pack for a selected sourced opportunity
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PrepareRoundRequest' }
      responses:
        '200':
          {
            description: Existing round,
            content: { application/json: { schema: { $ref: '#/components/schemas/Round' } } },
          }
        '201':
          {
            description: Round created,
            content: { application/json: { schema: { $ref: '#/components/schemas/Round' } } },
          }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /rounds/compare-offers:
    post:
      operationId: compareOffersRound
      summary: Commission one bounded comparison of complete owner supplied offer texts
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CompareOffersRoundRequest' }
      responses:
        '200':
          description: Existing intake and round for the exact request key
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CompareOffersRoundResponse' }
        '201':
          description: New intake and round
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CompareOffersRoundResponse' }
        '409': { $ref: '#/components/responses/Conflict' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /offer-comparisons/by-round/{roundId}:
    get:
      operationId: getOfferComparisonByRound
      summary: Read the requesting owner's saved comparison for one commissioned round
      security: [{ cookieAuth: [] }]
      parameters:
        - { name: roundId, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Saved comparison and captured qualitative selection status
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OfferComparisonResult' }
        '404': { $ref: '#/components/responses/NotFound' }
  /offer-comparisons/{id}:
    get:
      operationId: getOfferComparison
      summary: Read the requesting owner's immutable saved comparison
      security: [{ cookieAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Saved comparison and captured qualitative selection status
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OfferComparisonResult' }
        '404': { $ref: '#/components/responses/NotFound' }
  /rounds/process-input:
    post:
      operationId: processOwnerInput
      summary: Save one contextual owner input and commission its bounded record work
      description: The target identity and revision come from the visible record. An exact request key resumes the same saved input and round. A paused replacement requires the owner to name the exact paused round and revision.
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProcessInputRequest' }
      responses:
        '200':
          description: Existing saved input and round
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProcessInputResponse' }
        '201':
          description: Input saved and round commissioned
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProcessInputResponse' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '409': { $ref: '#/components/responses/Conflict' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /rounds/{id}:
    parameters: [{ $ref: '#/components/parameters/RecordId' }]
    get:
      operationId: getRound
      summary: Read a commissioned round
      security: [{ cookieAuth: [] }]
      responses:
        '200':
          {
            description: Round,
            content: { application/json: { schema: { $ref: '#/components/schemas/Round' } } },
          }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  parameters:
    RecordId:
      name: id
      in: path
      required: true
      schema: { type: string }
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: jobseek_session
      description: Owner session. Cookie mutations also require X-CSRF-Token and same-origin Origin.
    bearerAuth:
      type: http
      scheme: bearer
      description: Named agent token with operation-specific scopes.
  responses:
    Unavailable:
      description: Required round execution capability is unavailable
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    Conflict:
      description: Revision conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
...
    PrepareInterviewRequest:
      type: object
      additionalProperties: false
      required: [requestKey, opportunityId, context]
      properties:
        requestKey: { type: string, minLength: 1, maxLength: 200 }
        opportunityId: { type: string, minLength: 1 }
        context:
          {
            type: string,
            minLength: 1,
            maxLength: 30000,
            description: Complete owner supplied invitation and relevant context in one payload,
          }
...
    InterviewCommissionResponse:
      type: object
      required: [interviewId, round]
      properties:
        interviewId: { type: string }
        round: { $ref: '#/components/schemas/Round' }
...